Add Auto-Paste from Output Buffer
This commit adds auto-paste of the results of a run when the UI buffer
is hidden.

The mongosh output will be pasted on the line following the current
cursor position.

If the UI window is open, then we don't auto-paste, and the UI will
still be updated with the latest run results when toggling it open even
if they were auto-pasted.